+++ aws/doc/gen_paramsdoc.go
@@ -30,6 +30,9 @@ var generatedParamsDoc = map[string]map[string]string{
 	"attachinstance": {
 		"targetgroup": "The Amazon Resource Name (ARN) of the target group",
 	},
+	"attachinstanceprofile": {
+		"instance": "The ID of the instance",
+	},
 	"attachinternetgateway": {
 		"id":  "The ID of the Internet gateway",
 		"vpc": "The ID of the VPC",