+++ aws/transform.go
@@ -129,6 +129,8 @@ func initResource(source interface{}) (*graph.Resource, error) {
 		res = graph.InitResource(cloud.Policy, awssdk.StringValue(ss.PolicyId))
 	case *iam.AccessKeyMetadata:
 		res = graph.InitResource(cloud.AccessKey, awssdk.StringValue(ss.AccessKeyId))
+	case *iam.InstanceProfile:
+		res = graph.InitResource(cloud.InstanceProfile, awssdk.StringValue(ss.InstanceProfileId))
 	// S3
 	case *s3.Bucket:
 		res = graph.InitResource(cloud.Bucket, awssdk.StringValue(ss.Name))